Refactor GameCore classes into separate files
Mainly for readibilty. This will be the new convention
