+++ DropGrid.Client/DropGrid.Client.projitems
@@ -11,15 +11,21 @@
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)Asset\Asset.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Asset\AssetLoader.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ClientCore.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)GameEngine.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Asset\Sprite.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Asset\Spritesheet.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Asset\AssetRegistry.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Asset\AssetsToUse.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)State\GameState.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)State\LoadingState.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)State\MenuState.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)State\GameplayState.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)State\StateId.cs" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="$(MSBuildThisFileDirectory)Content\" />
     <Folder Include="$(MSBuildThisFileDirectory)Asset\" />
+    <Folder Include="$(MSBuildThisFileDirectory)State\" />
   </ItemGroup>
   <ItemGroup>
     <MonoGameContentReference Include="$(MSBuildThisFileDirectory)Content\Content.mgcb" />
