+++ DropGrid.Client/GameEngine.cs
@@ -123,104 +123,4 @@ namespace DropGrid.Client
             _currentState.OnEnter();
         }
     }
-
-    namespace State
-    {
-        /// <summary>
-        /// Each GameState handles one succinct set of game routines. They each have a unique identifiable state ID.
-        /// </summary>
-        abstract class GameState
-        {
-            public abstract StateId GetId();
-
-            public abstract void Draw(GameEngine engine, SpriteBatch spriteBatch, GameTime gameTime);
-            
-            public abstract void Update(GameEngine engine, GameTime gameTime);
-
-            public void OnEnter() { }
-            
-            public void OnExit() { }
-        }
-
-        /// <summary>
-        /// A list of supported states by the game client.
-        /// </summary>
-        public enum StateId
-        {
-            Initialise,
-            Menu,
-            Gameplay
-        }
-
-        /// <summary>
-        /// The initialisation state handles deferred content loading. It is too costly to load ALL game assets into memory during startup.
-        /// When new unloaded assets have been requested, we switch to this state and load them.
-        /// </summary>
-        class LoadingState : GameState
-        {
-            public override StateId GetId() => StateId.Initialise;
-
-            public override void Draw(GameEngine engine, SpriteBatch spriteBatch, GameTime gameTime)
-            {
-
-            }
-
-            public override void Update(GameEngine engine, GameTime gameTime)
-            {
-                if (IsInitializationFinished())
-                    engine.EnterState(StateId.Gameplay);
-            }
-
-            private bool IsInitializationFinished()
-            {
-                // TODO: Implement deferred resource loading later.
-                return true;
-            }
-        }
-
-        /// <summary>
-        /// Handles main menu logic.
-        /// </summary>
-        class MenuState : GameState
-        {
-            public override StateId GetId() => StateId.Menu;
-
-            public override void Draw(GameEngine engine, SpriteBatch spriteBatch, GameTime gameTime) => throw new NotImplementedException();
-
-            public override void Update(GameEngine engine, GameTime gameTime) => throw new NotImplementedException();
-        }
-
-        /// <summary>
-        /// The crux of the client-side game logic belongs here.
-        /// </summary>
-        class GameplayState : GameState
-        {
-            public override StateId GetId() => StateId.Gameplay;
-
-            // TODO: This is only temporary
-            bool loaded = false;
-            AssetsToUse assetsToUse;
-            public override void Draw(GameEngine engine, SpriteBatch spriteBatch, GameTime gameTime)
-            {
-                if (!loaded)
-                {
-                    assetsToUse = new AssetsToUse().Add(new Spritesheet("basic_ground_tiles", 128));
-
-                    AssetLoader.LoadQueue.Add(assetsToUse);
-                    AssetLoader.LoadQueue.LoadAll();
-                    loaded = true;
-                }
-                spriteBatch.Begin();
-                spriteBatch.Draw((Texture2D) ((Spritesheet) assetsToUse.GetAsset("basic_ground_tiles")).getSpriteAt(0, 0).GetData(), new Vector2(200, 200));
-                spriteBatch.End();
-            }
-
-            public override void Update(GameEngine engine, GameTime gameTime)
-            {
-                
-            }
-
-            
-        }
-    }
 }
