+++ DropGrid.Client/State/StateId.cs
@@ -1,0 +1,13 @@
+﻿using System;
+namespace DropGrid.Client.State
+{
+    /// <summary>
+    /// A list of supported states by the game client.
+    /// </summary>
+    public enum StateId
+    {
+        Initialise,
+        Menu,
+        Gameplay
+    }
+}
